+++ CHANGES.md
@@ -12,7 +12,8 @@ The released versions correspond to PyPI releases.
 * the default for `FakeFilesystem.shuffle_listdir_results` will change to `True` to reflect
   the real filesystem behavior
 
-## Unreleased
+## [Version 5.7.3](https://pypi.python.org/pypi/pyfakefs/5.7.3) (2024-12-15)
+Fixes a regression in version 5.7.3.
 
 ### Fixes
 * fixed a regression in version 5.7.2 that `tempfile` was not patched after pause/resume
@@ -22,7 +23,7 @@ The released versions correspond to PyPI releases.
 
 ### Infrastructure
 * run pytest-specific tests for all supported Python versions
-* pytest is only supported for versions >= 6.2.5, earlier version do not work
+* pytest is only supported for versions >= 6.2.5, earlier version do not work in Python >= 3.10
   due to a pytest issue - adapted tests and documentation
 
 ## [Version 5.7.2](https://pypi.python.org/pypi/pyfakefs/5.7.2) (2024-12-01)